Refactor rooms endpoint from `/api/adv/rooms` to `/api/rooms`
