+++ adventure/urls.py
@@ -5,5 +5,4 @@ urlpatterns = [
     url('init', api.initialize),
     url('move', api.move),
     url('say', api.say),
-    url('rooms', api.rooms),
 ]
